third_party/blink/web_tests/webmidi/requestmidiaccess-options-expected.txt

CONSOLE WARNING: line 58: Web MIDI will ask a permission to use even if the sysex is not specified in the MIDIOptions since M77, around September 2019. See https://www.chromestatus.com/feature/5138066234671104 for more details.
Test requestMIDIAccess with various options.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS navigator.requestMIDIAccess is defined.
PASS PermissionsHelper.setPermission is defined.


Start test with following configurations:
    options: {}
    configurations: {"midi":"denied","sysex":"denied"}
PASS permissions[permissionIndex].midi is 'denied'
Done.


Start test with following configurations:
    options: {"sysex":false}
    configurations: {"midi":"denied","sysex":"denied"}
PASS permissions[permissionIndex].midi is 'denied'
Done.


Start test with following configurations:
    options: {"sysex":true}
    configurations: {"midi":"denied","sysex":"denied"}
PASS permissions[permissionIndex].sysex is 'denied'
Done.


Start test with following configurations:
    options: {}
    configurations: {"midi":"granted","sysex":"denied"}
PASS permissions[permissionIndex].midi is 'granted'
Done.


Start test with following configurations:
    options: {"sysex":false}
    configurations: {"midi":"granted","sysex":"denied"}
PASS permissions[permissionIndex].midi is 'granted'
Done.


Start test with following configurations:
    options: {"sysex":true}
    configurations: {"midi":"granted","sysex":"denied"}
PASS permissions[permissionIndex].sysex is 'denied'
Done.


Start test with following configurations:
    options: {}
    configurations: {"midi":"granted","sysex":"granted"}
PASS permissions[permissionIndex].midi is 'granted'
Done.


Start test with following configurations:
    options: {"sysex":false}
    configurations: {"midi":"granted","sysex":"granted"}
PASS permissions[permissionIndex].midi is 'granted'
Done.


Start test with following configurations:
    options: {"sysex":true}
    configurations: {"midi":"granted","sysex":"granted"}
PASS permissions[permissionIndex].midi is 'granted'
PASS permissions[permissionIndex].sysex is 'granted'
PASS successfullyParsed is true

TEST COMPLETE
